register Test::DistManifest as a prereq and fail if it is not available
If Test::DistManifest is not installed, the generated test would just
skip itself. Since it is an author test, it shouldn't ignore errors like
this.

Register Test::DistManifest as a develop prereq, so that it can be
installed automatically by tooling, and make the test just use the
module directly rather than skipping on load failures.